backend/requirements.txt
fastapi
uvicorn
pinecone
python-dotenv
sentence-transformers
keybert
groq
